move test specific constructor to test
