+++ pom.xml
@@ -27,7 +27,7 @@
         <commons-cli.version>1.5.0</commons-cli.version>
         <xmlunit.version>2.9.1</xmlunit.version>
         <snakeyml.version>2.0</snakeyml.version>
-        <jimfs.version>1.3.0</jimfs.version
+        <jimfs.version>1.3.0</jimfs.version>
         <testExcludedGroups>HttpSmokeTest</testExcludedGroups>
     </properties>
