+++ src/main/java/digital/slovensko/autogram/server/dto/ServerSigningParameters.java
@@ -86,41 +86,6 @@ public class ServerSigningParameters {
                 identifier, checkPDFACompliance, getVisualizationWidth());
     }
 
-    public static ServerSigningParameters fromMap(Map<String, Object> map) {
-        var level = SignatureLevel.valueByName((String) map.get("level"));
-        var container = fromMapToEnum(ASiCContainerType.class, map.get("container"));
-        var containerFilename = (String) map.get("containerFilename");
-        var containerXmlns = (String) map.get("containerXmlns");
-        var packaging = fromMapToEnum(SignaturePackaging.class, map.get("packaging"));
-        var digestAlgorithm = fromMapToEnum(DigestAlgorithm.class, map.get("digestAlgorithm"));
-        var en319132 = (Boolean) map.get("en319132");
-        var infoCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class, map.get("infoCanonicalization"));
-        var propertiesCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class,
-                map.get("propertiesCanonicalization"));
-        var keyInfoCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class,
-                map.get("keyInfoCanonicalization"));
-        var schema = (String) map.get("schema");
-        var transformation = (String) map.get("transformation");
-        var identifier = (String) map.get("identifier");
-        var checkPDFACompliance = map.getOrDefault("checkPDFACompliance", "false") == "true";
-        var visualizationWidth = fromMapToEnum(VisualizationWidthEnum.class, map.get("visualizationWidth"));
-
-        return new ServerSigningParameters(
-                level,
-                container,
-                containerFilename,
-                containerXmlns,
-                packaging,
-                digestAlgorithm,
-                en319132,
-                infoCanonicalization,
-                propertiesCanonicalization,
-                keyInfoCanonicalization,
-                schema,
-                transformation,
-                identifier, checkPDFACompliance, visualizationWidth);
-    }
-
     private static <T extends Enum<T>> T fromMapToEnum(Class<T> clazz, Object obj) {
         var visualizationWidthStr = (String) obj;
         if (visualizationWidthStr == null)
