+++ src/test/java/digital/slovensko/autogram/SignHttpSmokeTest.java
@@ -10,8 +10,8 @@ import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
-import java.util.Map;
 import java.util.Base64;
+import java.util.Map;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.client.ClientProtocolException;
@@ -31,6 +31,12 @@ import org.yaml.snakeyaml.Yaml;
 
 import digital.slovensko.autogram.server.dto.Document;
 import digital.slovensko.autogram.server.dto.ServerSigningParameters;
+import digital.slovensko.autogram.server.dto.ServerSigningParameters.LocalCanonicalizationMethod;
+import digital.slovensko.autogram.server.dto.ServerSigningParameters.VisualizationWidthEnum;
+import eu.europa.esig.dss.enumerations.ASiCContainerType;
+import eu.europa.esig.dss.enumerations.DigestAlgorithm;
+import eu.europa.esig.dss.enumerations.SignatureLevel;
+import eu.europa.esig.dss.enumerations.SignaturePackaging;
 import digital.slovensko.autogram.server.dto.SignRequestBody;
 import digital.slovensko.autogram.server.dto.SignResponse;
 
@@ -101,9 +107,9 @@ public class SignHttpSmokeTest {
         signRequest.setEntity(new StringEntity(signRequestBody, "UTF-8"));
         var signResponse = clientBuilder.build().execute(signRequest);
         assertEquals(HttpStatus.SC_OK, signResponse.getStatusLine().getStatusCode());
-        System.out.println("Sign Response: " + signResponse.getStatusLine());
-        System.out.println("Sign Response: " + new String(
-                signResponse.getEntity().getContent().readAllBytes(), StandardCharsets.UTF_8));
+        // System.out.println("Sign Response: " + signResponse.getStatusLine());
+        // System.out.println("Sign Response: " + new String(
+        //         signResponse.getEntity().getContent().readAllBytes(), StandardCharsets.UTF_8));
     }
 
     @ParameterizedTest
@@ -123,7 +129,7 @@ public class SignHttpSmokeTest {
         var payloadMimeType = getNested(example, "payloadMimeType");
 
         var rDocument = new Document(document.get("filename"), document.get("content"));
-        var rParameters = ServerSigningParameters.fromMap((Map<String, Object>) parameters);
+        var rParameters = fromMap((Map<String, Object>) parameters);
         var rPayloadMimeType = (String) payloadMimeType;
 
         var body = new SignRequestBody(
@@ -163,4 +169,46 @@ public class SignHttpSmokeTest {
                 });
     }
 
+
+    private static ServerSigningParameters fromMap(Map<String, Object> map) {
+        var level = SignatureLevel.valueByName((String) map.get("level"));
+        var container = fromMapToEnum(ASiCContainerType.class, map.get("container"));
+        var containerFilename = (String) map.get("containerFilename");
+        var containerXmlns = (String) map.get("containerXmlns");
+        var packaging = fromMapToEnum(SignaturePackaging.class, map.get("packaging"));
+        var digestAlgorithm = fromMapToEnum(DigestAlgorithm.class, map.get("digestAlgorithm"));
+        var en319132 = (Boolean) map.get("en319132");
+        var infoCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class, map.get("infoCanonicalization"));
+        var propertiesCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class,
+                map.get("propertiesCanonicalization"));
+        var keyInfoCanonicalization = fromMapToEnum(LocalCanonicalizationMethod.class,
+                map.get("keyInfoCanonicalization"));
+        var schema = (String) map.get("schema");
+        var transformation = (String) map.get("transformation");
+        var identifier = (String) map.get("identifier");
+        var checkPDFACompliance = map.getOrDefault("checkPDFACompliance", "false") == "true";
+        var visualizationWidth = fromMapToEnum(VisualizationWidthEnum.class, map.get("visualizationWidth"));
+
+        return new ServerSigningParameters(
+                level,
+                container,
+                containerFilename,
+                containerXmlns,
+                packaging,
+                digestAlgorithm,
+                en319132,
+                infoCanonicalization,
+                propertiesCanonicalization,
+                keyInfoCanonicalization,
+                schema,
+                transformation,
+                identifier, checkPDFACompliance, visualizationWidth);
+    }
+    
+    private static <T extends Enum<T>> T fromMapToEnum(Class<T> clazz, Object obj) {
+        var visualizationWidthStr = (String) obj;
+        if (visualizationWidthStr == null)
+            return null;
+        return T.valueOf(clazz, visualizationWidthStr);
+    }
 }
